# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatnow API",
        "description": "Whatnow API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.whatnow.jazusoft.com/v1",
            "description": "API Base URL"
        }
    ],
    "paths": {
...
        "/org": {
            "get": {
                "tags": [
                    "Organisation"
                ],
                "summary": "Get all organisations (public)",
                "operationId": "04f86edf4b63c5b0f297e462633924b4",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/usage/applications": {
            "get": {
                "tags": [
                    "UsageLogs"
                ],
                "summary": "Get application usage logs",
                "operationId": "getApplicationLogs",
                "parameters": [
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Start date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "End date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/usage/endpoints": {
            "get": {
                "tags": [
                    "UsageLogs"
                ],
                "summary": "Get endpoint usage logs",
                "operationId": "getEndpointLogs",
                "parameters": [
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Start date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "End date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/usage/export": {
            "get": {
                "tags": [
                    "UsageLogs"
                ],
                "summary": "Export usage logs as CSV",
                "operationId": "exportUsageLogs",
                "parameters": [
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Start date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "End date for filtering logs (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
        "/usage/totals": {
            "get": {
                "tags": [
                    "UsageLogs"
                ],
                "summary": "Get usage log totals",
                "operationId": "getTotals",
                "parameters": [
                    {
                        "name": "society",
                        "in": "query",
                        "description": "Filter by society",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "subnational",
                        "in": "query",
                        "description": "Filter by subnational ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hazard",
                        "in": "query",
                        "description": "Filter by hazard type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by specific date (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Filter by language",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "deprecated": true,
                "security": []
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key"
            }
        }
    },
    "security": [
        {
            "BasicAuth": []
        }
    ]
}